Add ui tests to create event and change test srtucture
This add the unit test of the create event structure
Also, change the event tests structure and add the consftest.py module to allow share fixture to different tests modules


Restucture event tests

Moved all fixtures to conftest.py file
Put all test related file in a test directory
